wrote out more of the conversino handler
